Start refactoring platform detection
